noop rename: new space cards settings
